requirements.txt
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.3
openpyxl==3.1.2
tqdm==4.66.1 